#3797 comment: added security tests
